Feature iep3 (#21)
* Constexpr is implicitly inline

* Initial generic KML parser

- Parses gx:Track elements

* Provide abstract "KML track" class

* Add AbstractKMLParser

- Iterates over basic KML document structure
- FIX: proper double -> float conversion (waypoints store their
coordinates as float)
- FIX: flightradar24.com parser, parse Point (not Track)

* Only import first <gx:Track>

* Initial Flight Recorder CSV import

* Update flightradar24 and FlightAware names

- Add URL links to changelog

* Aircraft autocompleter now matches middle words

* Use C++ standard int types

* Parse engine and primary flight control data

* Import secondary flight control data

* FIX: Show spoiler position in percent

* Rename label to "Spoiler position"

* Import gear handles and light

* Use QFlags instead of reimplemented flag operations

- For convenience

+++ src/Plugin/src/Import/IGCImport/src/IGCParser.h
@@ -75,8 +75,8 @@ public:
               description(desc)
         {}
 
-        double latitude;
-        double longitude;
+        float latitude;
+        float longitude;
         QString description;
     } TaskItem;
